open-data-platform/requirements.txt
psycopg2-binary
Flask
Flask-SQLAlchemy
Flask-WTF
python-dotenv